+++ cedar-policy/CHANGELOG.md
@@ -12,6 +12,10 @@ Starting with version 3.2.4, changes marked with a star (*) are _language breaki
 
 ## [Unreleased]
 
+### Added
+
+- Public syntax tree representation for `Policy`, `Template` and `PolicySet` allowing programmatic manipulation of Cedar syntax (#816, #366).
+
 ## [4.10.0] - Coming soon
 
 Cedar Language Version: 4.5